[BE] FIX: CabinetExceptionHandler에 어울리는 메서드들 옮김
